"Add source-map-support types and update eslint rules"
This update adds '@types/source-map-support' dependency to 'package.json' and 'package-lock.json' to ensure better source map support in the application. The '.eslintrc.js' file has also been updated to comment out the 'node' environment and just minor changes made in the formatting of the rules section. These changes are done aiming to maintain high code quality and to enhance the debugging capabilities of the application.