Update database type in readme.md file
Readme.md file was updated to reflect a switch from SQLite to PostgreSQL as the base for data storage in the application. The reason for the change is to utilize PostgreSQL due to its robust functionality and wide acceptance. Also, a newline was added at the end of the file for better file formatting.