Add i18n support for translations
This commit adds international (i18n) localization support through the addition of 'nestjs-i18n' and 'class-validator' packages. These packages make it easier to manage multi-language content.

A new 'I18nTranslations' type was created to handle translations, and a 'testTranslation' query was added to the 'debug.resolver.ts' file to validate the new feature. Also, it introduces the translation files for English and Russian languages inside the 'i18n' directory.

The commit also includes changes to the 'app.module.ts' file to initialize the i18n module with the default language set to English, and it watches for changes in the 'i18n' directory. Nest-cli.json was also updated to include 'i18n' files as a part of the assets that need to be watched.

Lastly, 'class-validator' was added to validate input data and ensure the application is resilient against incorrect data.