Add GraphQL query complexity control
New module 'graphql-query-complexity' has been added to the package and a new folder 'complexity.plugin' has been introduced in common directory. 'ComplexityPlugin' has been added as a provider in 'app.module'. The changes have been implemented in order to manage and limit the complexity of GraphQL queries with a specified maximum limit.