Introduced class validator for resolvers, added rate limiter, used global pipes and enhanced Query complexity in the project.
This commit makes use of class-validation methods to improve the security and cleanliness of the code. Also, added rate limiters for /studio and other resolvers to prevent abuse of services. Global pipes are now used instead of local ones for uniform application (see `main.ts`). In addition, Query complexity is now properly monitored and warnings are issued whenever the complexity exceeds half of maximum limit. InputType classes are created for grouping the data inputs to have cleaner and manageable resolver parameters (see `auth.resolver.ts` and `types.ts`). Lastly, revisions were made on the todo.md, specifically on the tasks relating to the class validator, GraphQL extension and logger.