Add predb:gen script to package.json
A new script 'predb:gen' has been added to the package.json file. This script uses the rimraf package to remove the 'src/@generated' directory before generating a new one. It is necessary to clean up the old generated files to ensure the data structure of the application is correctly updated when there are database schema changes.